requirements.txt
alembic==1.13.1
amqp==5.2.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.6
limits==3.10.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
prompt-toolkit==3.0.43
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.7.1
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
